moved App.js into src, organized screens folder structure
