+++ app.json
@@ -4,9 +4,9 @@
     "slug": "mobile-app",
     "version": "1.0.0",
     "orientation": "portrait",
-    "icon": "assets/icon.png",
+    "icon": "src/assets/icon.png",
     "splash": {
-      "image": "assets/splash.png",
+      "image": "src/assets/splash.png",
       "resizeMode": "contain",
       "backgroundColor": "#ffffff"
     },
@@ -18,7 +18,7 @@
       "supportsTablet": true
     },
     "web": {
-      "favicon": "assets/favicon.png"
+      "favicon": "src/assets/favicon.png"
     }
   }
 }
