+++ package.json
@@ -18,10 +18,11 @@
     "expo": "^39.0.3",
     "expo-localization": "~9.0.0",
     "expo-status-bar": "~1.0.2",
-    "i18n-js": "^3.7.1",
+    "i18next": "^19.7.0",
     "lodash.memoize": "^4.1.2",
     "react": "16.13.1",
     "react-dom": "16.13.1",
+    "react-i18next": "^11.7.3",
     "react-native": "https://github.com/expo/react-native/archive/sdk-39.0.2.tar.gz",
     "react-native-gesture-handler": "^1.8.0",
     "react-native-onboarding-swiper": "^1.1.4",
