+++ src/assets/locales/de/translation.json
@@ -1,0 +1,4 @@
+{
+    "welcome": "Wilkommen",
+    "language" : "Sprache"
+}
