+++ src/assets/locales/en/translation.json
@@ -1,0 +1,4 @@
+{
+    "welcome": "Welcome",
+    "language": "Language"
+}
