+++ src/i18n.js
@@ -1,0 +1,31 @@
+import * as Localization from 'expo-localization';
+import i18n from "i18next";
+import { reactI18nextModule } from "react-i18next";
+
+import translationEN from "./assets/locales/en/translation.json";
+import translationDE from "./assets/locales/de/translation.json";
+
+const resources = {
+    en: {
+        translation: translationEN
+    },
+    de: {
+        translation: translationDE
+    }
+
+};
+
+i18n
+    .use(reactI18nextModule) //passes i18n down to react-i18next
+    .init({
+        lng: Localization.locale,
+        fallbackLng: 'en',
+        resources,
+        interpolation: {
+            escapeValue: false,
+        },
+        cleanCode: true,
+        
+    });
+
+export default i18n;
