+++ src/routes/AppNavigator.js
@@ -3,8 +3,11 @@ import React, {useState} from 'react';
 import BottomTabNavigator from './TabNavigator';
 import {OnboardingNavigator} from './StackNavigator';
 
+
+
+
 export default function AppNavigation() {
-  const [isOnboarded, setIsOnboarded] = useState(false);
+  const [isOnboarded, setIsOnboarded] = useState(true);
 
   if (!isOnboarded) {
     return <OnboardingNavigator />;
